@@ -1,6 +1,3 @@
-from pkg_resources import DistributionNotFound, get_distribution
+from importlib.metadata import version
 
-try:
-    __version__ = get_distribution("django-silk").version
-except DistributionNotFound:
-    pass
+__version__ = version("django-silk")
