@@ -1,5 +1,4 @@
 import logging
-import re
 import traceback
 
 from django.core.exceptions import EmptyResultSet
@@ -54,7 +53,7 @@ def _explain_query(connection, q, params):
 
         # currently we cannot use explain() method
         # for queries other than `select`
-        if re.match(f"^{prefix} .*", q):
+        if q.upper().startswith(prefix.upper()):
             # to avoid "EXPLAIN EXPLAIN", do not add prefix
             prefixed_query = q
         else:
